+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 0.1.0
+# 0.2.0 - 09/01/2012
+
+- Added ability to derive total count for search result (count prior to paging)
+- Resolves issue #2 https://github.com/morgan/kohana-paginate/issues/2
+- `Paginate::search` now a getter/setter for search query
+- All tests pass "OK (6 tests, 15 assertions)"
+
+# 0.1.0 - 06/02/2012
 
 - Initial release of Paginate
 - Support for Kohana Database, Kohana ORM and REST (via Dispatch)